Replaced flag with pflag, added multifile support
flag.Args() is supposed to provide only args that are positional, ie,
non-flag args. However the constraints on it make it basically
impossible to make that work without carefully constructed commandlines.
This adjusts the code to use the drop-in replacement pflag. Which has
the nice side effect of being POSIX-style flags.